==> churn/version2/0_churn_run_all.ipynb <==
#### Input

El input es el csv __Churn Brasil_plana mensualizada.csv__ que son los datos que nos pasan desde BI y tienen que estar ubicados en ./data/.

Esta notebook va a correr los cinco scripts que limpian los datos, entrena el modelo, calcula métricas, predice y calcula métricas de SHAP.

#### Output

Va a generar varios csvs en ./data. El más importante es __preds.csv__ que va a contener las predicciones. Para más detalles se puede entrar a cada una de las notebooks:

'1_churn_data_cleaning.ipynb'

'2_churn_model_train.ipynb'

'3_churn_metrics.ipynb'

'4_churn_prediction.ipynb'

'5_churn_SHAP_analysis.ipynb'

import nbformat
from nbconvert.preprocessors import ExecutePreprocessor

notebooks_list = ['1_churn_data_cleaning.ipynb',
                  '2_churn_model_train.ipynb',
                  '3_churn_metrics.ipynb',
                  '4_churn_prediction.ipynb',
                  '5_churn_SHAP_analysis.ipynb']

for notebook_filename in notebooks_list:
    with open(notebook_filename) as f:
        nb = nbformat.read(f, as_version=4)
        ep = ExecutePreprocessor(timeout=600, kernel_name='fast')
        _ = ep.preprocess(nb, {'metadata': {'path': './'}})
        print(notebook_filename + ' --- OK!')
