==> churn/version2/4_churn_prediction.ipynb <==
#### Input

Esta notebook va a levantar el csv __clientes_input.csv__ y va a hacer las predicciones para esos usuarios.

Las columnas deben ser como las del ejemplo __'./data/example_input.csv'__.

Si se quiere hacer una prueba para ver cómo funciona se puede cambiar el nombre del example_input.csv a clientes_input.csv y correr la notebook

#### Output

Va a generar el csv __preds.csv__.


import pandas as pd
import xgboost as xgb
from xgboost import XGBClassifier

!ls data

features = pd.read_csv('./data/features_validation.csv', index_col='idclientecrm')

### Modelo

model = XGBClassifier(learning_rate=0.1,n_estimators=100)
model.load_model("./data/models/xgb.model")

predicciones = pd.DataFrame(features.index).set_index('idclientecrm')

predicciones['pred'] = model.predict(features)
predicciones['weight'] = model.predict_proba(features).max(axis = 1)
predicciones = predicciones.sort_values(['pred','weight'], ascending = [True, False])

labels = pd.read_csv('./data/labels_validation.csv', index_col = 'idclientecrm')

df = pd.merge(predicciones, labels, left_index=True, right_index=True)

df['correct'] = (df['pred']==df['Status Cliente'])

len(df[df['correct']==True]) / len(df)

predicciones.to_csv('./data/preds.csv')
